Fix frontend tool definitions in LangGraph state (WIP) (#1053)
* Pre release mme/frontend-tool-docs

* update packages

* move convertJsonSchemaToZodSchema to shared and add test code

* exit pre

* Pre release mme/frontend-tool-docs

* implement convertActionsToDynamicStructuredTools

* Pre release mme/frontend-tool-docs

* pass frontend function to LangGraph JS

* remove console.log

* exit pre
