+++ CopilotKit/.changeset/afraid-teachers-attend.md
@@ -1,0 +1,18 @@
+---
+"@copilotkit/sdk-js": patch
+"next-openai": patch
+"next-pages-router": patch
+"node-express": patch
+"node-http": patch
+"@copilotkit/react-core": patch
+"@copilotkit/react-textarea": patch
+"@copilotkit/react-ui": patch
+"@copilotkit/runtime": patch
+"@copilotkit/runtime-client-gql": patch
+"@copilotkit/shared": patch
+"eslint-config-custom": patch
+"tailwind-config": patch
+"tsconfig": patch
+---
+
+lower case copilotkit property
