+++ CopilotKit/.changeset/big-mails-prove.md
@@ -1,0 +1,18 @@
+---
+"eslint-config-custom": patch
+"@copilotkit/runtime-client-gql": patch
+"next-pages-router": patch
+"tailwind-config": patch
+"@copilotkit/react-textarea": patch
+"node-express": patch
+"next-openai": patch
+"@copilotkit/react-core": patch
+"node-http": patch
+"tsconfig": patch
+"@copilotkit/react-ui": patch
+"@copilotkit/runtime": patch
+"@copilotkit/sdk-js": patch
+"@copilotkit/shared": patch
+---
+
+add zod conversion
