+++ examples/coagents-qa-native/agent-js/package.json
@@ -15,7 +15,7 @@
     "typescript": "^5.6.3"
   },
   "dependencies": {
-    "@copilotkit/sdk-js": "1.4.0",
+    "@copilotkit/sdk-js": "1.4.1-pre.2",
     "@langchain/anthropic": "^0.3.8",
     "@langchain/core": "^0.3.18",
     "@langchain/google-genai": "^0.1.4",
